refactor(arch): move test doubles out of src/main and eliminate GitLab4J leakage in TelegramServiceImpl
