+++ src/test/kotlin/net/raquezha/nuecagram/telegram/MockTelegramService.kt
